+++ packages/descriptions/src/index.tsx
@@ -478,6 +478,7 @@ const ProDescriptions = <
     actionRef,
     onRequestError,
     emptyText,
+    contentStyle,
     ...rest
   } = props;
 
@@ -614,6 +615,7 @@ const ProDescriptions = <
           {...rest}
           contentStyle={{
             minWidth: 0,
+            ...(contentStyle || {})
           }}
           extra={
             rest.extra ? (