+++ packages/layout/src/components/SiderMenu/index.tsx
@@ -19,6 +19,7 @@ const SiderMenuWrapper: React.FC<SiderMenuProps & PrivateSiderMenuProps> = (
     className,
     hide,
     prefixCls,
+    getContainer,
   } = props;
 
   const { token } = useContext(ProProvider);
@@ -63,6 +64,7 @@ const SiderMenuWrapper: React.FC<SiderMenuProps & PrivateSiderMenuProps> = (
         }}
         maskClosable
         closable={false}
+        getContainer={getContainer || false}
         width={siderWidth}
         styles={{
           body: {