+++ packages/provider/src/locale/zh_CN.tsx
@@ -46,7 +46,7 @@ export default {
     reload: '刷新',
     density: '密度',
     densityDefault: '正常',
-    densityLarger: '默认',
+    densityLarger: '宽松',
     densityMiddle: '中等',
     densitySmall: '紧凑',
   },